(issue 22) new database schema to code

+++ app/models/annotator.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Column, Integer, VARCHAR
+from sqlalchemy import Column, Integer, VARCHAR, BigInteger
 from sqlalchemy.orm import relationship
 
 from app import db
 
 
 class Annotator(db.Model):
-    id = Column(Integer, primary_key=True)
+    id = Column(BigInteger, primary_key=True)
     name = Column(VARCHAR(128), nullable=False, unique=True)
     annotations = relationship("Annotation", backref="annotator")
